+++ benchopt/benchmark.py
@@ -24,6 +24,10 @@ from .utils.shell_cmd import _run_shell_in_conda_env
 from .config import RAISE_INSTALL_ERROR
 
 
+# Global variable to access the benchmark currently running globally
+_RUNNING_BENCHMARK = None
+
+# Constant to name cache directory and folder of slurm outputs
 CACHE_DIR = '__cache__'
 SLURM_JOB_NAME = 'benchopt_run'
 
@@ -38,6 +42,11 @@ MISSING_DEPS_MSG = (
 )
 
 SUBSTITUTIONS = {"*": ".*"}
+
+
+def get_running_benchmark():
+    """Return the benchmark currently running."""
+    return _RUNNING_BENCHMARK
 
 
 class Benchmark:
@@ -65,6 +74,8 @@ class Benchmark:
     ):
         self.benchmark_dir = Path(benchmark_dir)
 
+        global _RUNNING_BENCHMARK
+        _RUNNING_BENCHMARK = self
         set_benchmark_module(self.benchmark_dir)
 
         # Load the benchmark metadat defined in `objective.py` or
@@ -99,6 +110,8 @@ class Benchmark:
             self.url = f"https://github.com/benchopt/{self.name}"
         else:
             self.name = Path(self.url).name
+        # replace dots to avoid issues with `with_suffix``
+        self.name = self.name.replace('.', '-')
 
     ####################################################################
     # Helpers to access and validate objective, solvers and datasets
@@ -333,8 +346,8 @@ class Benchmark:
         that are already in cache.
         """
 
-        # Create a cached function the computations in the benchmark folder
-        # and handle cases where we force the run.
+        # Create a cached version of `func` and handle cases where we force
+        # the run.
         func_cached = self.mem.cache(func, ignore=ignore)
         if force:
             assert not collect, "Cannot collect and force computation."