+++ benchopt/tests/test_run_config_files.py
@@ -33,7 +33,7 @@ CONFIG = """
 """
 
 
-def test_config_solver_no_params(no_debug_test):
+def test_config_solver_no_params(no_debug_log):
     with temp_benchmark(solvers=[TEST_SOLVER]) as benchmark:
         config_file = benchmark.benchmark_dir / "run_config.yml"
         config_file.write_text(CONFIG)
@@ -45,7 +45,7 @@ def test_config_solver_no_params(no_debug_test):
         out.check_output(r"RUN\(0\)", repetition=1)
 
 
-def test_config_solver_no_params_error(no_debug_test):
+def test_config_solver_no_params_error(no_debug_log):
     with temp_benchmark(solvers=[TEST_SOLVER]) as benchmark:
         config_file = benchmark.benchmark_dir / "run_config.yml"
         config_file.write_text(CONFIG.replace(" #PARAMS", "[param1=0]"))
@@ -63,7 +63,7 @@ def test_config_solver_no_params_error(no_debug_test):
                 ], standalone_mode=False)
 
 
-def test_config_solver_with_params(no_debug_test):
+def test_config_solver_with_params(no_debug_log):
     params = dict(param1=[None], param2=[None])
     with temp_benchmark(solvers=[
         TEST_SOLVER.replace("# PARAMETERS", f"parameters = {params}")
@@ -79,7 +79,7 @@ def test_config_solver_with_params(no_debug_test):
         out.check_output(r"solver1\[param1=None,param2=None\]", repetition=2)
 
 
-def test_config_solver_with_params_error(no_debug_test):
+def test_config_solver_with_params_error(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[param0=0]")
 
     params = dict(param1=[None], param2=[None])
@@ -102,7 +102,7 @@ def test_config_solver_with_params_error(no_debug_test):
                 ], standalone_mode=False)
 
 
-def test_config_solver_with_params_list_error(no_debug_test):
+def test_config_solver_with_params_list_error(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[0, 1]")
 
     params = dict(param1=[None], param2=[None])
@@ -121,7 +121,7 @@ def test_config_solver_with_params_list_error(no_debug_test):
                 ], standalone_mode=False)
 
 
-def test_config_solver_with_params_str_list(no_debug_test):
+def test_config_solver_with_params_str_list(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[param1=[0, 1]]")
 
     params = dict(param1=[None], param2=[None])
@@ -141,7 +141,7 @@ def test_config_solver_with_params_str_list(no_debug_test):
         out.check_output("param1=1,param2=None", repetition=2)
 
 
-def test_config_solver_with_params_str_value(no_debug_test):
+def test_config_solver_with_params_str_value(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[param1=0, param2=1]")
 
     params = dict(param1=[None], param2=[None])
@@ -161,7 +161,7 @@ def test_config_solver_with_params_str_value(no_debug_test):
         out.check_output("param1=0,param2=1", repetition=2)
 
 
-def test_config_solver_with_params_yaml_list(no_debug_test):
+def test_config_solver_with_params_yaml_list(no_debug_log):
     config = CONFIG.replace(" #PARAMS", ":\n            param1: [0, 1]")
 
     params = dict(param1=[None], param2=[None])
@@ -181,7 +181,7 @@ def test_config_solver_with_params_yaml_list(no_debug_test):
         out.check_output("param1=1,param2=None", repetition=2)
 
 
-def test_config_solver_with_params_yaml_items(no_debug_test):
+def test_config_solver_with_params_yaml_items(no_debug_log):
     config = CONFIG.replace(" #PARAMS", """:
               param1:
                 - 0
@@ -204,7 +204,7 @@ def test_config_solver_with_params_yaml_items(no_debug_test):
         out.check_output("param1=1", repetition=2)
 
 
-def test_config_solver_with_params_yaml_value(no_debug_test):
+def test_config_solver_with_params_yaml_value(no_debug_log):
     config = CONFIG.replace(" #PARAMS", """:
               param1: 0
               param2: 1
@@ -226,7 +226,7 @@ def test_config_solver_with_params_yaml_value(no_debug_test):
         out.check_output("param1=0,param2=1", repetition=2)
 
 
-def test_config_solver_with_params_complex_param(no_debug_test):
+def test_config_solver_with_params_complex_param(no_debug_log):
     config = CONFIG.replace(" #PARAMS", """:
               param1:
                 - test: 0
@@ -251,7 +251,7 @@ def test_config_solver_with_params_complex_param(no_debug_test):
         )
 
 
-def test_config_solver_with_one_params_list(no_debug_test):
+def test_config_solver_with_one_params_list(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[0, 1]")
 
     params = dict(param1=[None])
@@ -272,7 +272,7 @@ def test_config_solver_with_one_params_list(no_debug_test):
         out.check_output("param1=1", repetition=2)
 
 
-def test_config_solver_with_one_params_yaml_list(no_debug_test):
+def test_config_solver_with_one_params_yaml_list(no_debug_log):
     config = CONFIG.replace(" #PARAMS", ":\n          - 0\n          - 1")
 
     params = dict(param1=[None])
@@ -293,7 +293,7 @@ def test_config_solver_with_one_params_yaml_list(no_debug_test):
         out.check_output("param1=1", repetition=2)
 
 
-def test_config_solver_double_param_yaml_list(no_debug_test):
+def test_config_solver_double_param_yaml_list(no_debug_log):
     config = CONFIG.replace(
         " #PARAMS", ":\n            param1, param2: [[0, 1]]"
     )
@@ -316,7 +316,7 @@ def test_config_solver_double_param_yaml_list(no_debug_test):
         out.check_output("param1=0,param2=1", repetition=2)
 
 
-def test_config_solver_double_param_solver_yaml(no_debug_test):
+def test_config_solver_double_param_solver_yaml(no_debug_log):
     config = CONFIG.replace(" #PARAMS", "[param1=0]")
 
     params = {'param1, param2': [(None, None)]}