+++ doc/whats_new.rst
@@ -13,6 +13,9 @@ Version 1.6 - In development
 API
 ~~~
 
+- Add a `save_final_results` method to Objective. If implemented it is run after the last solver iteration, to get desired outputs to be saved to file system.
+  By `Pierre-Antoine Comby`_ (:gh:`722`)
+
 - Add native way to do cross-validation in a benchmark with
   ``Objective.cv`` attribute that change split for each repetition.
   By `Christopher Marouani`_ and `Thomas Moreau`_ (:gh:`623`).
@@ -20,6 +23,8 @@ API
 - Run-config files now support having parameters as nested dict, with
   potentially non-trivial structures (like dictionaries).
   By `Thomas Moreau`_ (:gh:`706`).
+
+- Raise error when an invalid install_cmd is provided. By `Jad Yehya` (:gh:`714`).
 
 CLI
 ~~~
@@ -30,11 +35,23 @@ CLI
 - Add ``--download`` option in ``benchopt install`` to allow downloading
   the data when installing the benchmark. By `Thomas Moreau`_ (:gh:`718`)
 
+- Add ``--no-timeout`` option in ``benchopt run`` to allow solvers to bypass
+  timeout. By `Célestin Eve`_ (:gh:`725`)
+
 FIX
 ~~~
 
+- Disable caching of diverged/errored runs. By `Julie Alberge`_ and
+  `Virginie Loison`_ (:gh:`735`)
+
 - Fix pickling of dynamic modules to allow for nested parallelism in
   distributed runs. By `Thomas Moreau`_ (:gh:`713`)
+
+DOC
+~~~
+
+- Add documentation for the `run_once` sampling strategy.
+  By `Mathieu Dagréou`_ (:gh:`700`).
 
 .. _changes_1_5_1:
 
@@ -155,7 +172,7 @@ PLOT
   the :class:`~benchopt.BaseSolver` class. By `Badr Moufad`_ (:gh:`543`)
 - Enable visualizing the objective as function of ``stopping_criterion``: time,
   iteration, or tolerance. By `Badr Moufad`_ (:gh:`479`)
-- Add button to share and set specific views on the plot. For now, the view need
+- Add button to share and set specific views on the plot. For now, the view needs
   to be defined manually in the benchmark config file but an export button will
   be added in follow up PRs.
   By `Tanguy Lefort`_, `Amélie Vernay`_ and `Thomas Moreau`_ (:gh:`552`).